+++ CMakeLists.txt
@@ -1,4 +1,4 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.10)
 PROJECT(camel-raisim-projects)
 
-add_subdirectory(camel-canine-single_leg)
+add_subdirectory(camel-canine-single_leg-left)
