+++ camel-canine-single_leg-left/CMakeLists.txt
@@ -1,4 +1,4 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.5)
-PROJECT(camel-canine-single_leg)
+PROJECT(camel-canine-single_leg-left)
 
 add_subdirectory(simulation)
