+++ camel-canine-single_leg-left/simulation/include/SingleLeggedSharedMemory.hpp
